refactor: adjust response on user creation
