+++ src/users/interfaces/user.interface.ts
@@ -1,15 +1,10 @@
 import { Document } from 'mongoose'
 
-export class User extends Document {
+export interface User extends Document {
   uuid: string
-
   name: string
-
   email: string
-
   password: string
-
   createdAt: Date
-
   updatedAt: Date
 }
