+++ src/users/interfaces/userResponse.interface.ts
@@ -1,0 +1,6 @@
+export interface UserResponse {
+  uuid: string
+  name: string
+  email: string
+  createdAt: Date
+}
